==> 01_data_load.ipynb <==
Upload the data to MongoDB. Remove duplicate recipes and cocktails then save the recipes as a dataframe and pickle the file to df_01.pkl.

# Imports

import numpy as np
import pandas as pd
import pickle as pkl
import matplotlib.pyplot as plt
%matplotlib inline

import json

from pymongo import MongoClient

# MongoDB

## Load JSON

with open('epicurious_recipes/full_format_recipes.json') as json_data:
    recipes_json = json.load(json_data)

## Connect to MongoDB Client

# ssh -NL 12345:localhost:27017 myaws

client = MongoClient(port=12345) # this is the port set by the SSH tunnel

address = client.address
db_names = client.database_names()
host = client.HOST
port = client.PORT

print('address: ', address, '\n',
      'databases: ', db_names, '\n',
      'host: ', host, '\n', 
      'port: ', port)

# assign Kojak database to db
db = client.kojak
db.collection_names()

# initialize new collection for recipe data within Kojak database
recipes = db.recipes

## Upload Recipes to kojak.recipe collection

# try inserting one document from the recipe json file to the recipe collection
recipes.insert_one(recipes_json[0])

# insert the rest of the documents from the recipe json file
for recipe in recipes_json[1:]:
    recipes.insert_one(recipe)

recipes.count(), df.shape

# Dataframe Creation

## Retrieve Data from kojak.recipe collection

# ssh -NL 12345:localhost:27017 myaws

client = MongoClient(port=12345) # this is the port set by the SSH tunnel

# assign Kojak database to db
db = client.kojak
db.collection_names()

recipe_db = db.recipes

df_mongo = pd.DataFrame(list(recipe_db.find()))
df_mongo.head()

## DF

df = df_mongo[['title', 'ingredients', 'directions', 'categories', 'desc']]
df.index.name = 'recipeID'
df.info()

## Drop duplicates and drink recipes

df.drop_duplicates(subset=['title'],keep='first',inplace=True)

df.dropna(axis=0, thresh=4,inplace=True)

df.reset_index(inplace=True)

cocktail_list = []
for i in range(len(df)):
    if 'Cocktail' in df.iloc[i].categories:
        cocktail_list.append(i)

df.drop(df.index[cocktail_list],inplace=True)

df.head()

df.drop(['level_0','index','recipeID'],axis=1,inplace=True)

df.reset_index(drop=True,inplace=True)

df.to_pickle('df_01.pkl')
