==> 07_custom_word2vec.ipynb <==
This notebook creates the custom word2vec model and the conitnuous bag of words dataframe for the recipes. First we take the original recipe dataframe with strings, preprocess the text into monograms and bigrams and check them against the Google W2V model's vocabulary to create a dictionary of words and two-word phrases for the custom W2V's corpus. That dictionary is used to train the MWETokenizer with acceptable bigrams that should always be bigrams. Next, we tokenize the recipes again, into sentences and words, and then train the custom word2vec model. Lastly, we process the recipes one more time to create continuous bags of words for the title, ingredients, directions, and recipes as a whole. That final dataframe is saved to be used in the next notebook to create document vectors.

# Imports

# Basics
import numpy as np
import pandas as pd
import math
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
import re
import string
import csv
from os import path
from PIL import Image
from wordcloud import WordCloud, STOPWORDS

# MongoDB
from pymongo import MongoClient

# natural language processing
import nltk
from nltk.util import ngrams
from textblob import TextBlob
from nltk import word_tokenize, pos_tag, ne_chunk
from nltk.corpus import stopwords
from ingreedypy import Ingreedy
stop_words = stopwords.words('english')

# sklearn
from sklearn.feature_extraction import text 
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer

from sklearn.preprocessing import Normalizer
from sklearn.preprocessing import scale
from sklearn.utils import shuffle
from sklearn.model_selection import train_test_split
from sklearn.cluster import KMeans

from sklearn.decomposition import PCA
from sklearn.decomposition import TruncatedSVD

from sklearn import metrics
from sklearn.metrics import pairwise_distances
from sklearn.metrics import silhouette_score

# gensim
from gensim import corpora, models, similarities, matutils
import gensim

# Other
import warnings
warnings.filterwarnings('ignore')


# Google WV

import gensim
google_vec_file = '/Users/carliebadder/Downloads/GoogleNews-vectors-negative300.bin'
g_model = gensim.models.KeyedVectors.load_word2vec_format(google_vec_file, binary=True)

# MWETokenizer with bi- and tri-grams

# text data
df_text = pd.read_pickle('df_01.pkl')

df_text.title

def preprocess_ngrams(text):
    text = text.lower()
    doc = word_tokenize(text)
#     print('token',doc)
    doc2 = []
    for word in doc:
        if '-' in word:
            two_words = word.split('-')
            doc2.extend(two_words)
        elif '/' in word:
            two_words = word.split('/')
            doc2.extend(two_words)
        else:
            doc2.append(word)
#     print('\-',doc2)

    doc2 = [re.sub('[%s]' % re.escape(string.punctuation), '', word) for word in doc2]
    doc2 = [word for word in doc2 if word.isalpha()]
    monograms = doc2 #[word for word in doc2 if word not in stop_words]
#     bigrams = []
#     for i in range(len(monograms)-1):
#         bigrams.append(monograms[:i] + [monograms[i]+'_'+monograms[i+1]] + monograms[i+2:])
    bigrams = list(ngrams(monograms,2))
    bigrams = [w for w in bigrams if '_'.join(w) in g_model]
#     trigrams = list(ngrams(monograms,3))
#     trigrams = [w for w in trigrams if '_'.join(w) in model]
    
#     print('mono',monograms)
#     print('bi',(bigrams))
#     print('tri',(trigrams))
    
    return bigrams #+trigrams

df_text.title[7293], preprocess_ngrams(df_text.title[7293])

title_ngrams = []
# Titles
for i in range(len(df_text)):
    temp_word_list = preprocess_ngrams(df_text.title[i])
    title_ngrams.extend(temp_word_list)

ing_ngrams = []
# Ingredients
for i in range(len(df_text)):
    try:
        temp_ing_list = df_text.ingredients[i]
        for ing in temp_ing_list:
            temp_word_list = preprocess_ngrams(ing)
            ing_ngrams.extend(temp_word_list)
    except:
        continue

dir_ngrams = []
# Directions
for i in range(len(df_text)):
    try:
        temp_dir_list = df_text.directions[i]
        for dir_ in temp_dir_list:
            temp_word_list = preprocess_ngrams(dir_)
            dir_ngrams.extend(temp_word_list)
    except:
        continue

desc_ngrams = []
# Descriptions
for i in range(len(df_text)):
    try:
        temp_desc_list = df_text.desc[i]
        for desc in temp_desc_list:
            temp_word_list = preprocess_ngrams(desc)
            desc_ngrams.extend(temp_word_list)
    except:
        continue

ngram_list = title_ngrams + ing_ngrams + dir_ngrams + desc_ngrams

(ngram_list[590])

('lemon','zest') in ngram_list

from nltk.tokenize import MWETokenizer

tokenizer = MWETokenizer(ngram_list)
tokenizer.tokenize('This is a test lemon peel'.split())
    

import pickle
with open('ngram_list.pkl', 'wb') as f:
    pickle.dump(ngram_list, f)

# Epicurious Corpus for Word2Vec Model

pd.options.display.max_colwidth = 500
df_text.head()

def preprocess(text):
    text = text.lower()
    doc = text.split()
#     print('token',doc)
    doc2 = []
    for word in doc:
        if '-' in word:
            two_words = word.split('-')
            doc2.extend(two_words)
        elif '/' in word:
            two_words = word.split('/')
            doc2.extend(two_words)
        else:
            doc2.append(word)

    doc2 = [re.sub('[%s]' % re.escape(string.punctuation), '', word) for word in doc2]
    doc2 = [word for word in doc2 if word.isalpha()]
#     print(doc2)
    monograms = tokenizer.tokenize(doc2)
    
    return monograms

preprocess(df_text.title[7913]) # df_text.title[7913], 

df_text.ingredients[333][0], preprocess(df_text.ingredients[333][0])

# Titles
title_sent = []

for i in range(len(df_text)):
    temp_word_list = preprocess(df_text.title[i])
    title_sent.append(temp_word_list)

# Ingredients
ing_sent = []

for i in range(len(df_text)):
    try:
        temp_ing_list = df_text.ingredients[i]
        for ing in temp_ing_list:
            temp_word_list = preprocess(ing)
            ing_sent.append(temp_word_list)
    except:
        continue

# Directions
dir_sent = []

for i in range(len(df_text)):
    try:
        temp_dir_list = df_text.directions[i]
        for dir_ in temp_dir_list:
            temp_word_list = preprocess(dir_)
            dir_sent.append(temp_word_list)
    except:
        continue

# Descriptions
desc_sent = []

for i in range(len(df_text)):
    try:
        temp_desc_list = df_text.desc[i]
        for desc in temp_desc_list:
            temp_word_list = preprocess(desc)
            desc_sent.append(temp_word_list)
    except:
        continue

epicurious_texts = title_sent + ing_sent + dir_sent + desc_sent

with open('epicurious_texts.pkl', 'wb') as f:
    pickle.dump(epicurious_texts, f)

# Word2Vec

model = gensim.models.Word2Vec(epicurious_texts, size=200, window=5, min_count=1, workers=2, sg=1)

model.similarity('lemon_zest','lemon_peel')

# CBOW

def preprocess_cbow(text):
    text = text.lower()
    doc = text.split()
#     print('token',doc)
    doc2 = []
    for word in doc:
        if '-' in word:
            two_words = word.split('-')
            doc2.extend(two_words)
        elif '/' in word:
            two_words = word.split('/')
            doc2.extend(two_words)
        else:
            doc2.append(word)

    doc2 = [re.sub('[%s]' % re.escape(string.punctuation), '', word) for word in doc2]
    doc2 = [word for word in doc2 if word.isalpha()]
    doc2 = [word for word in doc2 if word not in stop_words]
    doc2 = [word for word in doc2 if word not in measurements] # only use for ingredients
#     print(doc2)
    monograms = tokenizer.tokenize(doc2)
    
    return monograms

# Titles
title_cbow = title_sent

measurements = ['cup','cups','C','c','gram','grams','g','kilogram','kilograms','kg','liter','liters','L','l',
               'pound','pounds','lb','milliliter','milliliters','ml','mL','ounce','ounces','oz','pint','pints','pt',
               'teaspoon','teaspoons','t','tsp','tablespoon','tablespoons','T','TB','Tbl','Tbsp','tbsp','quart','quarts','qt',
               'dash','pinch','piece','pieces','slice','slices','sheet','sheets','log','stick','sticks','head',
               'chopped','diced','sliced','cored','seeded','trimmed','chunk','chunks','minced','divided','cut',
               'packed','fresh','large','medium','small','sprig','thin','thinly','thick','inch','well','chilled',
               'half','halved','halves','peeled','crumbled','crushed','dice','packaged','purchased','cubes','package',
               'separated','fine','finely','coarsely','freshly','generous','diameter','stemmed','packages','thaw',
               'thawed','additional','grated','grate', 'pitted','rounds','wedges','skinned','deveined','shredded',
               'unpeeled','optional','preferably','long','frozen','room','temperature','quartered','sometimes','called',
               'reserve','reserved','plus','lengthwise','parts','discarded','parts','bunch','shelled','available',
               'seasonally','markets','bottle','canned','drained','ingredient','ground','whole','round','style',
               'cans','leaves','strips','ground','bag','strip','split','supermarket','supermarkets','specialty',
               'crosswise','size','ring','rings','square','squares','colors','julienne','minute','minutes','dry',
               'chopping','shell','notes','shells','wash','diagonal','accompaniment','string','strings','press',
               'dish','firm','dash','dashes']

# split? supermarket(s), specialty, vanilla_extract?, crosswise, matchstick, size, rings, squares, colors, julienne,
# minute(s), shells, shell, notes, diagonal, wash, dry, chopping, accompaniment, string(s), dish, press, firm,
# dash, dashes

# Ingredients
ing_cbow = []
for i in range(len(df_text)):
    try:
        temp_ing_list = df_text.ingredients[i]
        temp_word_list = []
        for ing in temp_ing_list:
#             print([w for w in (ing.split()) if w not in measurements])
#             temp_ing = [w for w in (ing.split()) if w not in measurements]
#             temp_ing = ' '.join(temp_ing)
#             print(temp_ing)
            temp_word_list.append(preprocess_cbow(ing))
        ing_cbow.append(temp_word_list)
    except:
        continue

ing_cbow[:4]

# Directions
dir_cbow = []

for i in range(len(df_text)):
    try:
        temp_dir_list = df_text.directions[i]
        temp_word_list = []
        for dir_ in temp_dir_list:
            temp_word_list.append(preprocess_cbow(dir_))
        dir_cbow.append(temp_word_list)
    except:
        continue

# Descriptions
desc_cbow = []

for i in range(len(df_text)):
    try:
        temp_desc_list = df_text.desc[i]
#         temp_word_list = []
#         for desc in temp_desc_list:
#             print(desc)
        desc_cbow.append(preprocess_cbow(temp_desc_list))
#         desc_cbow.append(temp_word_list)
    except:
        desc_cbow.append([])

len(desc_cbow), len(dir_cbow)

df_cbow = pd.DataFrame(columns=['title','ingredients','directions','desc'])

df_cbow['title'] = title_cbow
df_cbow['ingredients'] = ing_cbow
df_cbow['directions'] = dir_cbow
df_cbow['desc'] = desc_cbow

df_cbow.to_pickle('df_07_2.pkl')
