==> 08_custom_doc2vec.ipynb <==
# Imports

# Basics
import numpy as np
import pandas as pd
import math
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
import re
import string
import csv
from os import path
from PIL import Image
from wordcloud import WordCloud, STOPWORDS
import pickle as pkl

# MongoDB
from pymongo import MongoClient

# natural language processing
import nltk
from textblob import TextBlob
from nltk import word_tokenize, pos_tag, ne_chunk
from nltk.corpus import stopwords
from ingreedypy import Ingreedy

# sklearn
from sklearn.feature_extraction import text 
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer

from sklearn.preprocessing import Normalizer
from sklearn.preprocessing import scale
from sklearn.utils import shuffle
from sklearn.model_selection import train_test_split
from sklearn.cluster import KMeans

from sklearn.decomposition import PCA
from sklearn.decomposition import TruncatedSVD

from sklearn import metrics
from sklearn.metrics import pairwise_distances
from sklearn.metrics import silhouette_score

# gensim
from gensim import corpora, models, similarities, matutils

# Other
import warnings
warnings.filterwarnings('ignore')


# Data from 02_cbow_process

df = pd.read_pickle('df_07_2.pkl')

df.head()

df.ingredients[0]

# Word2Vec Custom Model

import gensim

with open('epicurious_texts.pkl', 'rb') as f:
    epicurious_texts = pkl.load(f)

model = gensim.models.Word2Vec(epicurious_texts, size=200, window=5, min_count=1, workers=2, sg=1)

def document_vector(word2vec_model, doc):
    # remove out-of-vocabulary words
    doc = [word for word in doc if word in model.wv]
#     print(doc)
    return np.mean(word2vec_model[doc], axis=0)

measurements = ['cup','cups','C','c','gram','grams','g','kilogram','kilograms','kg','liter','liters','L','l',
               'pound','pounds','lb','milliliter','milliliters','ml','mL','ounce','ounces','oz','pint','pints','pt',
               'teaspoon','teaspoons','t','tsp','tablespoon','tablespoons','T','TB','Tbl','Tbsp','tbsp','quart','quarts','qt',
               'dash','pinch','piece','pieces','slice','slices','sheet','sheets','log','stick','sticks']

## Create Document Vectors

# get vector for each ingredient in ingredients
# then take the mean to get the ingredient list vector
ingredient_vectors = []
for i in range(len(df)): # each recipe
    ing_list = df.ingredients[i] # the list of ingredients
#     print(ing_list)
    ing_vec_temp = []
    for j in range(len(ing_list)): # each item in the list of ingredients
        try:
            ing_vec_temp.append(document_vector(model, ing_list[j])) # create individual ingredient vector
        except:
            pass
#     print(ing_vec_temp)
    ingredient_vectors.append(ing_vec_temp) # list of lists of individual ingredient vectors
    

ingredient_list_vectors = []
for i in range(len(ingredient_vectors)):
    try:
        ing_vstack = np.vstack(ingredient_vectors[i])
        ing_mean = np.mean(ing_vstack, axis=0)
    except:
        ing_mean = float('NaN')
    ingredient_list_vectors.append(ing_mean)

# get vector for each direction item in directions
# then take the mean to get the direction list vector
direction_vectors = []
for i in range(len(df)): # each recipe
    dir_list = df.directions[i] # the list of directions
    dir_vec_temp = []
    for j in range(len(dir_list)): # each item in the list of directions
        try:
            dir_vec_temp.append(document_vector(model, dir_list[j])) # create individual direction item vector
        except:
            pass
    direction_vectors.append(dir_vec_temp) # list of lists of individual direction item vectors

direction_list_vectors = []
for i in range(len(direction_vectors)):
    try:
        dir_vstack = np.vstack(direction_vectors[i])
        dir_mean = np.mean(dir_vstack, axis=0)
    except:
        dir_mean = float('NaN')
    direction_list_vectors.append(dir_mean)

# get vector for each title
title_vectors = []
for i in range(len(df)): # each recipe
    title_temp = df.title[i] # the title of the recipe
    try:
        title_vectors.append(document_vector(model, title_temp)) # list of title vectors
    except:
        title_vectors.append(float('NaN'))


# take the mean of the title, ingredient list, and direction list vectors 
# to get the recipe vector
recipe_vectors = []
for i in range(len(df)):
    doc_temp = [title_vectors[i], direction_list_vectors[i], ingredient_list_vectors[i]]
    doc_temp = [x for x in doc_temp if not np.isnan(x).all()]
    doc_temp = np.vstack(doc_temp)
    recipe_vectors.append(np.mean(doc_temp, axis=0))


(recipe_vectors[0].shape)

# df_03

df_03 = pd.DataFrame(columns=['title','ingredient_ind','ingredient_avg','directions_ind','directions_avg','recipe_vector'])

df_03['title'] = title_vectors
df_03['ingredient_ind'] = ingredient_vectors
df_03['ingredient_avg'] = ingredient_list_vectors
df_03['directions_ind'] = direction_vectors
df_03['directions_avg'] = direction_list_vectors
df_03['recipe_vector'] = recipe_vectors

df_03.head()

df_03.to_pickle('df_07_2_vecs.pkl')
